misc: Ignore TS, JS and HTML files with stylelint
This is almost useless if we consider how the stylelint command is
executed (i.e. only on (S)CSS files). However, some editors (e.g. Vim)
can check the syntax in other files and may display errors in the
editor.

This change ensure that CSS is not checked in TS, JS and HTML files by
the editors.