a11y: Fix focus after navigation
Angular doesn't bring the focus back to the top of the page after
clicking on a navigation link. Also, since the page isn't reloaded, the
title isn't announced by screen readers.

To fix both issues, I added a listener on route events to give the focus
to the top of the page on changes. The page title is announced thanks to
the `aria-label` set on the DOM which has the focus.