a11y: Add quick access to the main content
Angular breaks normal usage of links and I can't use a simple
`href="#content-focus"` link.

The Angular `fragment` directive has a strange behaviour: if the
fragment is already included in the current URL, nothing happens.

The only solution left to me was to create a component method to give
the focus to the main content.