tec: Allow to override the backendUrl
This commit reverts 72180f26d7f879c3786a98fbde5b48c64c38b21c, but is
more flexible: if the configuration file doesn't exist, it will fallback
to the `/api` endpoint which is handled by the proxy config.

This allows administrators to bypass an eventual proxy configuration and
gain in performance.