fix: Load types in Settings only when connected
The types API is only available to connected users, but we were calling
it from the login page. The request was failing and the login form was
inaccessible.

Now, the types are loaded in two cases: during application
initialization if the user is connected, and after a successful login
(otherwise the types would not be loaded after login except if the user
refreshes the page).