tec: Refactor the API
The idea behind this refactoring is to have a file which implements
directly the backend API (api/v1.ts). Other files extend this base with
methods which have more sense from an application point of view.

This will help to implement API calls for other types and split the code
to not have one big file for the whole API.